requirements.txt
beautifulsoup4==4.9.1
bootstrap4==0.1.0
Django==2.2.15
django-bootstrap4==2.2.0
django-chartjs==2.2.1
importlib-metadata==1.7.0
pkg-resources==0.0.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
